style: :art: rearrange to separate functionalities
also update README
